feat: Add dynamic status icon and participant list to HostPage
This commit modifies the HostPage widget to include a dynamic status icon and a list of participants. The HostPage widget is now a StatefulWidget to enable state changes. The status icon toggles between a check mark and an error icon when the "締め切り" button is pressed. The list of participants is displayed using a ListView.builder and each participant is represented by a ListTile.

The changes also include updating the onPressed event handler for the "開催者ページ" button in the MyHomePage widget to navigate to the HostPage with the conference name and a list of participants.

This commit addresses the need to display the status and participants in the HostPage and improves the user experience by providing dynamic functionality.