feat: Update icon color in HostPage
The code changes update the `HostPage` widget in `main.dart` to change the color of the `Icon` component. The `leading` property of the `ListTile` now includes the `color` parameter, which is set to the `iconColor` variable. This change was made to improve the visual appearance of the page.

Note: The recent user commits and repository commits were reviewed, but they did not provide any additional relevant information for the commit message.